requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
Django==3.0.8
django-admin-rangefilter==0.6.1
django-cors-headers==3.4.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
mysqlclient==2.0.1
openpyxl==3.0.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
pyfcm==1.4.7
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
